% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_labels.R
\name{available_labels}
\alias{available_labels}
\title{A list of the available Alaska area labels (900 +!)}
\usage{
available_labels()
}
\value{
A dataframe with 1 column of label names
}
\description{
The available area labels in alphabetical order
}
