% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_shapefiles.R
\name{available_shapefiles}
\alias{available_shapefiles}
\title{A list of the available shapefiles from MACEReports}
\usage{
available_shapefiles()
}
\value{
prints the available shapefiles from MACEReports
}
\description{
A list of the available shapefiles from MACEReports
}
