% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_position.R
\name{break_position}
\alias{break_position}
\title{break_position}
\usage{
break_position(x)
}
\arguments{
\item{x}{a factor with levels at which breaks will be specified}
}
\description{
function needed to put lines below each reporting region- identify break positions. from:
https://stackoverflow.com/questions/59263028/categorize-and-highlight-table-sections-with-flextable
}
\keyword{internal}
