% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_biomass_at_length_table_summer_goa.R
\name{build_biomass_at_length_table_summer_goa}
\alias{build_biomass_at_length_table_summer_goa}
\title{Create biomass at length table for MACE summer GOA cruise reports}
\usage{
build_biomass_at_length_table_summer_goa(biomass_nums_length_data)
}
\arguments{
\item{biomass_nums_length_data}{The name of the dataframe that contains the pollock
biomass at length data; this dataframe is created in the cruise report markdown
process using the function \code{get_biomass_and_nums_data_function.R}; this provides all
the data that is then used to create a 'current survey only/pollock only' dataframe named
\code{current_survey_pollock_biomass_nums}. This dataframe is the easiest input to
\code{build_numbers_at_length_table_summer_goa} function}
}
\value{
A list with two items: item 1 is the Flextable table object, item 2 is the table caption.
}
\description{
Build the pollock biomass-at-length table used in summer GOA reports. This will report all lengths
from 10-70 cm, and will do so for each Report Number (i.e. survey region) that was defined for the analysis.
If you have an especially small (<10 cm) or large (>70 cm) length class, the table will report these sizes as opposed
to the nomimal 10 cm- 70 cm range.
}
\examples{
\dontrun{
# build the table and caption
biomass_table_list <-
  build_biomass_at_length_table_summer_goa
(biomass_nums_length_data <- current_survey_pollock_biomass_nums)

# pull out table and caption from the list
biomass_table <- biomass_table_list[[1]]
biomass_caption <- biomass_table_list[[2]]
}
}
\author{
Mike Levine
}
