% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_haul_table.R
\name{build_haul_table}
\alias{build_haul_table}
\title{Create trawl haul effort table for MACE cruise reports}
\usage{
build_haul_table(haul_data)
}
\arguments{
\item{haul_data}{The name of the dataframe that contains all needed values; this dataframe is created
in the cruise report markdown process using the get_haul_table_data.R function. \code{haul_data} is
checked to make sure that all columns and data types are present and correct; if there are problems with
\code{haul_data}, errors will be returned that (hopefully) point out where problems exist.}
}
\value{
A list with two items: item 1 is the Flextable table object, item 2 is the table caption.
}
\description{
Produce a haul table for use in standard MACE cruise reports.
}
\examples{
\dontrun{
#build the table and caption
haul_table_list = build_haul_table(haul_data = shelikof_haul_data)

#pull out table and caption from the list
haul_table = haul_table_list[[1]]
haul_caption = haul_table_list[[2]]
}

}
\author{
Mike Levine
}
