% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_numbers_at_age_table_summer_goa.R
\name{build_numbers_at_age_table_summer_goa}
\alias{build_numbers_at_age_table_summer_goa}
\title{build_numbers_at_age_table_summer_goa}
\usage{
build_numbers_at_age_table_summer_goa(biomass_nums_age_data)
}
\arguments{
\item{biomass_nums_age_data}{The name of the dataframe that contains the pollock numbers-at-age data;
this dataframe is created in the cruise report markdown process using the function
\code{get_biomass_and_nums_age_data_function.R}; this provides all the data that is used to create
a 'current survey only/pollock only' dataframe named \code{current_survey_pollock_biomass_nums_age}.
This dataframe is the easiest input to the \code{build_numbers_at_age_table_summer_goa} function.}
}
\value{
A list with two items: item 1 is the Flextable table object, item 2 is the table caption.
}
\description{
Build the pollock numbers-at-age table used in summer GOA reports. This will report numbers at
age (millions of fish) for ages from 1- maximum reported age in the survey,
and will do so for each report number (i.e. survey region) that was defined for the analysis.
}
\examples{
\dontrun{
#build the table and caption
numbers_at_age_table_list =
 build_numbers_at_age_table_summer_goa
 (biomass_nums_age_data = current_survey_pollock_biomass_nums_age)
numbers_at_age_table =  numbers_at_age_table_list[[1]]
numbers_at_age_caption = numbers_at_age_table_list[[2]]
}
}
\author{
Mike Levine
}
