% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_numbers_at_length_table_winter_goa.R
\name{build_numbers_at_length_table_winter_goa}
\alias{build_numbers_at_length_table_winter_goa}
\title{Create numbers at length table for MACE winter cruise reports}
\usage{
build_numbers_at_length_table_winter_goa(biomass_nums_length_data, region_name)
}
\arguments{
\item{biomass_nums_length_data}{The name of the dataframe that contains all the abundance-at-length values;
this dataframe is created in the cruise report markdown process using the function
\code{get_total_biomass_by_length.R}; this provides all the data that is then used to create a
dataframe named \code{shelikof_sel_corr_surveys_totals_by_length_and_region}. This dataframe is the easiest input to
\code{build_numbers_at_length_table_winter_goa} function.}

\item{region_name}{The name of the reporting region that you want to include in this table
(generally "Shelikof Strait" in winter GOA reports)}
}
\value{
A list with two items: item 1 is the Flextable table object, item 2 is the table caption.
}
\description{
Build the pollock numbers-at-length table used in winter GOA reports. The data used here is usually
limited to the 'selectivity-corrected era' within Shelikof Strait (i.e. 2008-current). This function will report all
lengths from 5-76 cm, and will do so for any region that was defined for the analysis.
If you have an especially small (<5 cm) or large (>76 cm) length class, the table will report these sizes as opposed
to the nomimal 5 cm- 76 cm range.
}
\examples{
\dontrun{
# build the table and the caption
nums_table_list <- build_numbers_at_length_table_winter_goa(
  biomass_nums_length_data = shelikof_sel_corr_surveys_totals_by_length_and_region,
  region_name = "Shelikof Strait"
)

# pull out the table and caption from the list
nums_length_table <- nums_table_list[[1]]
nums_length_caption <- nums_table_list[[2]]
}
}
\author{
Mike Levine
}
