% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_specimen_table_summer_goa.R
\name{build_specimen_table_summer_goa}
\alias{build_specimen_table_summer_goa}
\title{Create biological specimen table for MACE summer GOA cruise reports}
\usage{
build_specimen_table_summer_goa(specimen_data, scaling_hauls_list = NULL)
}
\arguments{
\item{specimen_data}{a dataframe returned from the 'get_specimen_table_data.R' function, summer GOA cruise reports}

\item{scaling_hauls_list}{an (optional) numeric vector of hauls that were used to scale backscatter in your analysis.
These hauls are returned from the \code{scaling_hauls} dataframe in summer GOA cruise reports.
If not specified, all hauls in the survey will be reported in the table.#'}
}
\value{
A list with two items: item 1 is the Flextable table object, item 2 is the table caption.
}
\description{
Produce the specimen table for use in standard MACE GOA summer reports. This differs from the specimen table in other MACE reports by including the GOA reporting species- currently pollock, capelin, and POP.
}
\examples{
\dontrun{
#build the table and caption
specimen_table_list = build_specimen_table_summer_goa(specimen_data = specimen_table_data,
                                                      scaling_hauls_list = scaling_hauls$EVENT_ID)

#unpack tables and captions
specimen_table = specimen_table_list[[1]]
specimen_caption = specimen_table_list[[2]]
}
}
\author{
Mike Levine
}
