% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stick_legend.R
\name{build_stick_legend}
\alias{build_stick_legend}
\title{Return a legend that can be added to stickplots}
\usage{
build_stick_legend(
  stick_data,
  legend_pos = "right",
  legend_color = "white",
  legend_label = NULL,
  plot_expansion = NULL,
  fontsize = 11,
  barwidth = 0.75
)
}
\arguments{
\item{stick_data}{A 'stick' dataframe returned from \code{build_sf_sticks}.}

\item{legend_pos}{One of 'right' or 'left'; legend bar will be placed at the corresponding corner.}

\item{legend_color}{Default is white for standard MACE basemaps; specify alternates in standard \code{ggplot}-compatible formats}

\item{legend_label}{Defaults to a label in standard MACE units (t/nmi^2); specify alternate if needed to reflect your units}

\item{plot_expansion}{This controls the amount that a legend gets 'pushed' by (default is 0, which works well with for most uses). Set from 0-1. If you are working with the MACEReports::get_basemap_layers to get a basemap and you specify a large plot expansion (> .1) It works well to set this argument to the basemap plot expansion/2.}

\item{fontsize}{The font size for the stickplot legend. Defaults to a reasonable size for reports; make larger for presentations. ~18 works.}

\item{barwidth}{The width of the legend bar. Defaults to a reasonable size for reports; make larger for presentations. ~ 3 works.}
}
\description{
Accepts a dataframe returned from the \code{build_sf_sticks} function and returns a nicely placed legend bar.This legend is half the height of the tallest value in the dataframe. This object can be plotted using \code{ggplot2::geom_sf()} or base \code{plot}.
}
\examples{

library(ggplot2)
library(sf)

dat <- data.frame(
  "x" = c(-152.2, -150.3, -159.4),
  "y" = c(55.2, 55.8, 55.6),
  "z" = c(75000, 400000, 280000),
  "species" = c("a", "a", "b")
)
sticks = build_sf_sticks(x = dat$x, y = dat$y, z = dat$z, rotation = 15, crs = 3338)
legend = build_stick_legend(stick_data = sticks, legend_pos = 'left', legend_color = 'black')
ggplot()+
  geom_sf(data = sticks)+
  legend

}
