% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_df.R
\name{check_input_df}
\alias{check_input_df}
\title{check_input_df}
\usage{
check_input_df(template_df, input_df)
}
\arguments{
\item{template_df}{A dataframe that contains all the required columns and formats for the target table. This
should be in the \code{data} folder and documented in the \code{R} folder.}

\item{input_df}{The current survey data that you want to generate a table for.}
}
\value{
Nothing, if successful. Returns (hopefully) informative errors if dataframe isn't in the right format.
}
\description{
Checks that dataframes required by functions are complete and contain the proper class types.
}
\author{
Mike Levine
}
\keyword{internal}
