% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_plot_limits.R
\name{easy_plot_limits}
\alias{easy_plot_limits}
\title{Set your ggplot map limits a bit more easily}
\usage{
easy_plot_limits(plot_limits_data, plot_expansion = 0.05)
}
\arguments{
\item{plot_limits_data}{A \code{sf} spatial dataframe; this is required and used to map extent. Use the same dataframe as you used for \code{get_basemap_layers}!}

\item{plot_expansion}{This controls the amount of buffer around the basemap (default is 5\% (0.05) around the extent of the data; set from 0-1). This should match the value used for \code{get_basemap_layers}.}
}
\value{
a \code{ggplot2} object that will limit the extent of your ggplot map.
}
\description{
Returns a \code{ggplot2} object that will limit the extent of your ggplot map to cover only the data you care about. This will be slightly larger than the extent of \code{plot_limits_data}; users can still use \code{ggplot2::coord_sf} to fine-tune the plot extent.
}
\examples{
\dontrun{
library(ggplot2)
library(sf)
library(MACEReports)

dat <- data.frame(
  "x" = c(-152.2, -150.3, -159.4),
  "y" = c(55.2, 55.8, 55.6),
  "z" = c(7500, 40000, 28000),
  "species" = c("a", "a", "b")
)

# create an sf dataframe
dat <- sf::st_as_sf(dat, coords = c("x", "y"), crs = 4326)

# convert CRS to a reasonable projection
dat <- sf::st_transform(dat, crs = "EPSG:3338")

# return the basemap as a ggplot layer
mace_basemap <- get_basemap_layers(plot_limits_data = dat)

# now go get some other bathymetry
}
}
