% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basemap_layers.R
\name{get_basemap_layers}
\alias{get_basemap_layers}
\title{Create a MACE-themed basemap in the Bering Sea}
\usage{
get_basemap_layers(
  plot_limits_data,
  bathy = TRUE,
  contours = NULL,
  management_regions = NULL,
  SSL_critical_habitat = NULL,
  alaska_3nmi_buffer = NULL,
  land_fill_color = "#616161",
  land_outline_color = "black",
  region = NULL
)
}
\arguments{
\item{plot_limits_data}{A \code{sf} spatial dataframe; this is required and used to define the base map extent and projection.}

\item{bathy}{By default, a bathymetric baselayer based on the GEBCO (https://www.gebco.net/) gridded bathymetric dataset
is included in the basemap; If \code{FALSE}, bathymetric baselayer will not be included}

\item{contours}{Provide contour lines at requested depths. Specify depths as positive numeric values,
i.e. \code{c(200,100,50)}}

\item{management_regions}{If \code{TRUE}, will add NMFS management regions to basemap}

\item{SSL_critical_habitat}{If \code{TRUE}, will add Steller Sea Lion critical habitat buffers to basemap}

\item{alaska_3nmi_buffer}{If \code{TRUE}, will add the ADFG 3 nmi management buffer to basemap}

\item{land_fill_color}{If you'd like a different fill color on landmasses, specify as required by \code{ggplot2}.}

\item{land_outline_color}{If you'd like a different outline color on landmasses, specify as required by \code{ggplot2}.}

\item{region}{If you are working in the GOA- selecting 'goa' will give you a higher-resolution bathymetry layer.}
}
\value{
A list of class \code{ggplot} containing information required for plotting a basemap.
}
\description{
Returns a base map. This map is returned as a ggplot2 object that more complex maps can be built on top of.
It provides land, bathymetry, and, optionally, a variety of common layers including the NMFS management areas, 3 NMI buffer regions,
and Steller Sea Lion exclusions.
This layer will be slightly larger than the extent of \code{plot_limits_data}; users should still use \code{ggplot2::coord_sf}
to fine-tune the plot extent.
}
\examples{
\dontrun{
library(ggplot2)
library(sf)
library(MACEReports)

dat = data.frame('x' = c(-152.2, -150.3, -159.4),
'y' = c(55.2, 55.8, 55.6),
'z' = c(7500,40000, 28000),
'species' = c('a', 'a', 'b'))

#create an sf dataframe
dat = sf::st_as_sf(dat, coords = c('x', 'y'), crs = 4326)
#convert CRS to a reasonable projection
dat = sf::st_transform(dat, crs = 'EPSG:3338')

#return the basemap as a ggplot layer
mace_basemap = get_basemap_layers(plot_limits_data = dat)

#note that this basemap will replace the usual call to ggplot() as the start of your plot:
mace_basemap+
 geom_sf(data = dat, aes(color = species))

#You can add other commonly used MACE layers to the basemap
mace_basemap = get_basemap_layers(plot_limits_data = dat, management_regions = TRUE)
mace_basemap

mace_basemap = get_basemap_layers(plot_limits_data = dat, alaska_3nmi_buffer = TRUE)
mace_basemap

mace_basemap = get_basemap_layers(plot_limits_data = dat, SSL_critical_habitat = TRUE)
mace_basemap

mace_basemap = get_basemap_layers(plot_limits_data = dat, bathy = FALSE, contours = c(200, 100, 50))
mace_basemap
}

}
\author{
Mike Levine
}
