% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interpolated_color_bins.R
\name{get_interpolated_color_bins}
\alias{get_interpolated_color_bins}
\title{Return logical breaks that can be used to plot abundance values}
\usage{
get_interpolated_color_bins(z, n_bins = 9, style = "log10", nice_labels = TRUE)
}
\arguments{
\item{z}{a vector of abundance values}

\item{n_bins}{the number of bins you'd like to return; default is 9. This is ignored if style 'log10'.}

\item{style}{the algorithm used in binning; default is 'log10', which places abundance values in log10 bins; 'kmeans' also produces good results.}

\item{nice_labels}{return label values in linear (vs log) space that are formatted for use on plots; this is meant to be #' used in conjunction with values generated by \code{get_interpolated_plot_vals}, which return log10 units.}
}
\description{
Return binned abundance data for plotting interpolated values with greater contrast. This is a wrapper around
the \link[classInt:classIntervals]{classInt::classIntervals} function to make it easier to return binned color values.
}
\author{
Mike Levine
}
