% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefile.R
\name{get_shapefile}
\alias{get_shapefile}
\title{Return shapefiles that are used in MACEReports}
\usage{
get_shapefile(shapefile_name, crs = "EPSG:3338")
}
\arguments{
\item{shapefile_name}{The name of a shapefile to access; current list is available with \code{available_shapefiles()}}

\item{crs}{The coordinate reference system (CRS) into which shapefiles will be projected. By default, objects will be returned using the default CRS Albers Equal Area Alaska, (EPSG 3338). Any other valid projection can be specified. Note that this relies on the \link[sf:st_crs]{sf::st_crs} function- make sure you are using a valid CRS argument.}
}
\description{
Return layers as \code{sf} objects (shapefiles) or xyz dataframes (rasters).
}
\examples{
#return Alaska land in default CRS
ak_land = get_shapefile(shapefile_name = 'alaska_land')

#return Alaska land in a different CRS
ak_land = get_shapefile(shapefile_name = 'alaska_land', crs = 'EPSG:4326')
}
