% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{globals}
\alias{globals}
\title{globals definitions
define the 'undefined' variables in the package to silence the devtools::check() warning
no visible binding for global variable'...}
\description{
globals definitions
define the 'undefined' variables in the package to silence the devtools::check() warning
no visible binding for global variable'...
}
