% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass_nums_by_length.R
\name{plot_biomass_nums_by_length}
\alias{plot_biomass_nums_by_length}
\title{Get biomass- and numbers- by length plots with dual y-axis}
\usage{
plot_biomass_nums_by_length(
  length_vector,
  biomass_vector,
  numbers_vector,
  add_totals = FALSE
)
}
\arguments{
\item{length_vector}{A vector of fish lengths (units are assumed to be cm in 1 cm increments). This vector must be equal in length to \code{biomass_vector} and \code{numbers_vector}.}

\item{biomass_vector}{A vector of fish weights (units are assumed to be KG). This must should be equal in length to \code{length_vector} and \code{numbers_vector}.}

\item{numbers_vector}{A vector of fish numbers (units are assumed to be individual fish). This must should be equal in length to \code{length_vector} and \code{biomass_vector}.}

\item{add_totals}{Option to add the totals as the title of the plot. Default is to not add a title. Set to TRUE if you'd like a title with the total biomass and number values.}
}
\value{
A ggplot plot object. Note that the actual plot is a R baseplot; it is returned as a ggplot object so that it can be more easily integrated into Rmarkdown documents.
}
\description{
Returns a dual-axis plot with numbers as bars and biomass as lines; each abundance measure is plotted on independent y-axis scales. Plot units are automatically scaled to millions of fish OR thousands of fish/ 1000s of tons OR tons based on the magnitude of the biomass (if there is greater than 1 million KG at any length, the larger scale is used).
}
\examples{
# generate a simulated fish population
lengths <- round(rnorm(mean = 30, sd =3, n = 100000), digits = 0)
weights <- lengths^3 * 0.01
numbers <- sample(c(0, 3e3), size = 100000, replace = TRUE)

# plot it
plot_biomass_nums_by_length(length_vector = lengths,
 biomass_vector = weights,
  numbers_vector = numbers)

# plot it with totals added as text
plot_biomass_nums_by_length(length_vector = lengths,
biomass_vector = weights,
numbers_vector = numbers,
add_totals = TRUE)

}
