% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight_at_length.R
\name{plot_weight_at_length}
\alias{plot_weight_at_length}
\title{Plot mean fish weight at length for the current survey, with historical weight at length context.}
\usage{
plot_weight_at_length(
  survey_vector,
  length_vector,
  weight_vector,
  current_survey
)
}
\arguments{
\item{survey_vector}{a vector identifying the survey a given length and weight value is from. This vector must be equal in length to \code{length_vector} and \code{weight_vector}.}

\item{length_vector}{A vector of fish lengths (units are assumed to be cm). This vector must be equal in length to \code{survey_vector} and \code{weight_vector}.}

\item{weight_vector}{A vector of fish weights (units are assumed to be kg). This vector must be equal in length to \code{survey_vector} and \code{length_vector}.}

\item{current_survey}{Identify which survey is the current survey so that it can be highlighted in plots (and removed from the historical comparison).}
}
\value{
A list with 3 items:
Item 1 = A ggplot object
Item 2 = A dataframe of the current survey weight at length
Item 3 = A dataframe summarizing the historical survey weight at length
}
\description{
Present the current weight at length (mean +/- 1 s.d., with lengths binned at 1 cm increments) with historical survey observations plotted as the range of past observations as well the historical mean +/- 1 s.d. To be consistent with the MACE standard approach, only classes with >= 5 fish are plotted; in cases with < 5 fish the estimated weight at length based on a linear regression is used instead (see De Robertis and Williams 2008: Weight–Length Relationships in Fisheries Studies: The Standard Allometric Model Should Be Applied with Caution.).
}
\examples{
\dontrun{
#load example data
library(MACEReports)
data("pollock_length_weight_age_data")

#gather plot and summary dataframes
pollock_wt_len <- plot_weight_at_length(
survey_vector = pollock_length_weight_age_data$SURVEY,
 length_vector = pollock_length_weight_age_data$FORK_LENGTH,
 weight_vector = pollock_length_weight_age_data$ORGANISM_WEIGHT,
  current_survey = 202104)

# show the plot
pollock_wt_len[[1]]
# get the current survey summary
current_survey_pollock_wt_len <- pollock_wt_len[[2]]
# and the historical survey summary
historical_survey_pollock_wt_len <- pollock_wt_len[[3]]
}
}
