% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelikof_biomass.R
\docType{data}
\name{shelikof_biomass}
\alias{shelikof_biomass}
\title{shelikof_biomass}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 126647 rows and 7 columns.
}
\usage{
shelikof_biomass
}
\description{
An example dataset consisting of pollock biomass and biomass/nmi^2 values at length on a per-interval basis.
}
\details{
#' @format A data frame with 126647 rows and 6 columns:
\describe{
  \item{INTERVAL}{A unique identifier for each 0.5 nmi acoustic interval (numeric)}
  \item{LENGTH}{pollock length (cm; numeric)}
  \item{START_LATITUDE}{Latitude (decimal degrees; numeric)}
  \item{START_LONGITUDE}{Longitude (decimal degrees; numeric)}
  \item{BIOMASS}{Pollock biomass in interval, scaled by interval width (kg; numeric)}
  \item{BIOMASS_NM2}{Pollock biomass in interval, density (kg/nmi-2; numeric)}
  \item{year}{survey year (numeric)}
}
#' @source Historic MACE reported values
}
\keyword{datasets}
