% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelikof_historic_numbers_by_age_1981_to_2007.R
\docType{data}
\name{shelikof_historic_numbers_by_age_1981_to_2007}
\alias{shelikof_historic_numbers_by_age_1981_to_2007}
\title{shelikof_historic_numbers_by_age_1981_to_2007}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 19 rows and 25 columns.
}
\source{
Dataset initially compiled from historic cruise report tables
}
\usage{
shelikof_historic_numbers_by_age_1981_to_2007
}
\description{
values are stored as static, formatted 'canned' values gleaned from historic cruise reports.
If you want to change the static dataframe (for example, removing years as more selectivity-corrected values become
available, the data is available within MACEReports package: see \code{shelikof_historic_numbers_by_age_1981_to_2007}.
}
\keyword{datasets}
