% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelikof_sst.R
\docType{data}
\name{shelikof_sst}
\alias{shelikof_sst}
\title{An example dataset consisting of sea surface temperatures.}
\format{
A data frame with 12840 rows and 3 columns:
\describe{
  \item{LATITUDE}{Latitude (decimal degrees; numeric)}
  \item{LONGITUDE}{Longitude (decimal degrees; numeric)}
  \item{MEASUREMENT_VALUE}{Sea surface temperature (degrees C; numeric)}
}
#' @source Historic MACE reported values
}
\usage{
shelikof_sst
}
\description{
An example dataset consisting of sea surface temperatures.
}
\keyword{datasets}
