% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_inputs.R
\name{stick_inputs}
\alias{stick_inputs}
\title{stick_inputs}
\usage{
stick_inputs(x, y, z, group_variable = NULL)
}
\arguments{
\item{x}{x- position}

\item{y}{Y- position}

\item{z}{Abundance value (the value that will be used to scale sticks)}

\item{group_variable}{If specified, will specify the group for each stick; this can be used
to identify sticks by group.}
}
\description{
An internal function to check the inputs to build_sf_sticks and stop if there are issues
}
\examples{
\dontrun{
 dat = data.frame('x' = c(-152.2, -150.3, -159.4),
'y' = c(55.2, 55.8, 55.6),
'z' = c(75000,400000, 280000))

stick_inputs(x, y ,z)
}
}
\keyword{internal}
