% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_nums_format.R
\name{table_nums_format}
\alias{table_nums_format}
\title{table_nums_format}
\usage{
table_nums_format(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
a character for each value.
}
\description{
An internal function to format catch weights and percents and numbers percents conditionally as
we've generally presented them in tables. Specifically, if values = NA, label as '-',
If values are <=0.01, label as '<0.01', otherwise, format the value as having a single decimal point for display
}
\examples{
\dontrun{

values = c(1000, NA, .001)

 formatted_values = table_nums_format(values)

}
}
\keyword{internal}
