% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_df_catch_data_summer_goa.R
\docType{data}
\name{template_df_catch_data_summer_goa}
\alias{template_df_catch_data_summer_goa}
\title{template_df_catch_data_summer_goa}
\format{
A data frame with 5 rows and 8 columns:
\describe{
  \item{EVENT_ID}{Event number (numeric)}
  \item{GEAR}{Gear type (character)}
  \item{SPECIES_CODE}{RACE species code (numeric)}
  \item{SUBCATEGORY}{Species subcategory (character)}
  \item{SCIENTIFIC_NAME}{species scientific name (character)}
  \item{COMMON_NAME}{species common name (character)}
  \item{WEIGHT_IN_HAUL}{species catch weight (kg; numeric)}
  \item{NUMBER_IN_HAUL}{species catch number (individuals; numeric)}
}
}
\source{
Template dateset was gathered using get_raw_catch_summary.R function in Summer 2021 GOA Cruise Report
}
\usage{
template_df_catch_data_summer_goa
}
\description{
A subset of catch data, with columns and formats required for creating the standard catch tables
used in Summer Gulf of Alaska cruise reports
}
\keyword{datasets}
