% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/template_df_numbers_biomass_at_length_tables_winter_goa.R
\docType{data}
\name{template_df_numbers_biomass_at_length_tables_winter_goa}
\alias{template_df_numbers_biomass_at_length_tables_winter_goa}
\title{template_df_numbers_biomass_at_length_tables_winter_goa}
\format{
A data frame with 5 rows and 8 columns:
\describe{
  \item{SURVEY}{Survey number (numeric)}
  \item{SURVEY_YEAR}{Survey year (numeric)}
  \item{SPECIES_CODE}{RACE species code (numeric)}
  \item{TIMESERIES}{Timeseries associated with report region (character)}
  \item{REPORT_REGION}{Report region associated with REPORT_NUMBER (character)}
  \item{LENGTH}{fork length(cm; numeric)}
  \item{BIOMASS}{Biomass (kg; numeric)}
  \item{NUMBERS}{Numbers of individuals (numeric)}
}
}
\source{
Template dataset was gathered using winter 2025 GOA Cruise Report
}
\usage{
template_df_numbers_biomass_at_length_tables_winter_goa
}
\description{
A subset of abundance data, with columns and formats required for creating the standard biomass- and numbers-
by length tables used in winter Gulf of Alaska cruise reports.
}
\keyword{internal}
