% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_catch_table_summer_goa.R
\name{build_catch_table_summer_goa}
\alias{build_catch_table_summer_goa}
\title{build_catch_table_summer_goa}
\usage{
build_catch_table_summer_goa(
  catch_data,
  specimen_data,
  scaling_hauls_list = NULL
)
}
\arguments{
\item{catch_data}{a dataframe returned from the 'get_catch_table_data.R' function, summer GOA cruise reports}

\item{specimen_data}{a dataframe returned from the 'get_specimen_table_data.R' function, summer GOA cruise reports}

\item{scaling_hauls_list}{an (optional) numeric vector of hauls that were used to scale backscatter in your analysis.
These hauls are returned from the \code{scaling_hauls} dataframe in summer GOA cruise reports.
If not specified, all hauls in the survey will be reported in tables.}
}
\value{
A list (or multiple lists, if there are multiple gear types)
with two items: item 1 is the Flextable table object, item 2 is the table caption.
}
\description{
build the catch by species and numbers by gear; create a table that is in the format
used in the Gulf of Alaska summer reports. This differs from the winter version in the format of the input data.
}
\author{
Mike Levine
}
