% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stick_legend.R
\name{build_stick_legend}
\alias{build_stick_legend}
\title{build_stick_legend}
\usage{
build_stick_legend(stick_data, legend_pos = "right")
}
\arguments{
\item{stick_data}{A 'stick' dataframe returned from \code{build_sf_sticks}.}

\item{legend_pos}{One of 'right' or 'left'; legend bar will be placed at the corresponding corner.}
}
\description{
Accepts a dataframe returned from the \code{build_sf_sticks} function and returns a nicely placed legend bar.
This legend is half the height of the tallest value in the dataframe.
This object can be plotted using \code{ggplot2::geom_sf()} or base \code{plot}.
}
\examples{

library(ggplot2)
library(sf)

dat = data.frame('x' = c(-152.2, -150.3, -159.4),
'y' = c(55.2, 55.8, 55.6),
'z' = c(75000,400000, 280000),
'species' = c('a', 'a', 'b'))
sticks = build_sf_sticks(x = dat$x, y = dat$y, z = dat$z, rotation = 15, crs = 3338)
legend = build_stick_legend(stick_data = sticks, legend_pos = 'left')
ggplot()+
  geom_sf(data = sticks)+
  geom_sf(data = legend, color = 'red')

}
