% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{globals}
\alias{globals}
\title{globals definitions}
\description{
globals definitions
}
\keyword{'undefined'}
\keyword{binding}
\keyword{define}
\keyword{devtools::check()}
\keyword{for}
\keyword{global}
\keyword{in}
\keyword{internal}
\keyword{no}
\keyword{package}
\keyword{silence}
\keyword{the}
\keyword{to}
\keyword{variable'...}
\keyword{variables}
\keyword{visible}
\keyword{warning}
